moved public class HashParameters to new file
